+++ extensions/broker-server/src/test/java/de/sovity/edc/ext/brokerserver/services/api/CatalogApiTest.java
@@ -31,8 +31,7 @@ import de.sovity.edc.ext.brokerserver.db.jooq.Tables;
 import de.sovity.edc.ext.brokerserver.db.jooq.enums.ConnectorContractOffersExceeded;
 import de.sovity.edc.ext.brokerserver.db.jooq.enums.ConnectorDataOffersExceeded;
 import de.sovity.edc.ext.brokerserver.db.jooq.enums.ConnectorOnlineStatus;
-import de.sovity.edc.ext.wrapper.api.common.mappers.utils.AssetJsonLdUtils;
-import de.sovity.edc.utils.jsonld.vocab.Prop;
+import de.sovity.edc.ext.wrapper.api.common.model.UiAssetCreateRequest;
 import jakarta.json.JsonObject;
 import lombok.SneakyThrows;
 import org.eclipse.edc.junit.annotations.ApiTest;
@@ -67,9 +66,9 @@ class CatalogApiTest {
     @BeforeEach
     void setUp(EdcExtension extension) {
         extension.setConfiguration(createConfiguration(TEST_DATABASE, Map.of(
-                BrokerServerExtension.CATALOG_PAGE_PAGE_SIZE, "10",
-                BrokerServerExtension.DEFAULT_CONNECTOR_DATASPACE, "MDS",
-                BrokerServerExtension.KNOWN_DATASPACE_CONNECTORS, "Example1=https://my-connector2/api/dsp,Example2=https://my-connector3/api/dsp"
+            BrokerServerExtension.CATALOG_PAGE_PAGE_SIZE, "10",
+            BrokerServerExtension.DEFAULT_CONNECTOR_DATASPACE, "MDS",
+            BrokerServerExtension.KNOWN_DATASPACE_CONNECTORS, "Example1=https://my-connector2/api/dsp,Example2=https://my-connector3/api/dsp"
         )));
     }
 
@@ -79,7 +78,12 @@ class CatalogApiTest {
             // arrange
             var today = OffsetDateTime.now().withNano(0);
 
-            var assetJsonLd = getAssetJsonLd("my-asset", Map.of(Prop.Dcterms.TITLE, "My Asset"));
+            var assetJsonLd = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset")
+                    .title("My Asset")
+                    .build()
+            );
 
             // Dataspace: MDS
             createConnector(dsl, today, "https://my-connector/api/dsp");
@@ -91,7 +95,7 @@ class CatalogApiTest {
 
             var query = new CatalogPageQuery();
             query.setFilter(new CnfFilterValue(List.of(
-                    new CnfFilterValueAttribute("dataSpace", List.of("Example1"))
+                new CnfFilterValueAttribute("dataSpace", List.of("Example1"))
             )));
 
             var result = brokerServerClient().brokerServerApi().catalogPage(query);
@@ -108,7 +112,12 @@ class CatalogApiTest {
             // arrange
             var today = OffsetDateTime.now().withNano(0);
 
-            var assetJsonLd = getAssetJsonLd("my-asset", Map.of(Prop.Dcterms.TITLE, "My Asset"));
+            var assetJsonLd = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset")
+                    .title("My Asset")
+                    .build()
+            );
 
             createConnector(dsl, today, "https://my-connector/api/dsp");
             createDataOffer(dsl, today, "https://my-connector/api/dsp", assetJsonLd);
@@ -118,7 +127,7 @@ class CatalogApiTest {
 
             var query = new CatalogPageQuery();
             query.setFilter(new CnfFilterValue(List.of(
-                    new CnfFilterValueAttribute("connectorEndpoint", List.of("https://my-connector/api/dsp"))
+                new CnfFilterValueAttribute("connectorEndpoint", List.of("https://my-connector/api/dsp"))
             )));
 
             var result = brokerServerClient().brokerServerApi().catalogPage(query);
@@ -151,9 +160,9 @@ class CatalogApiTest {
             // assert that the filter values are correct
             var dataSpace = getAvailableFilter(result, "dataSpace");
             assertThat(dataSpace.getValues()).containsExactly(
-                    new CnfFilterItem("Example1", "Example1"),
-                    new CnfFilterItem("Example2", "Example2"),
-                    new CnfFilterItem("MDS", "MDS")
+                new CnfFilterItem("Example1", "Example1"),
+                new CnfFilterItem("Example2", "Example2"),
+                new CnfFilterItem("MDS", "MDS")
             );
         });
     }
@@ -164,9 +173,12 @@ class CatalogApiTest {
             // arrange
             var today = OffsetDateTime.now().withNano(0);
 
-            var assetJsonLd = getAssetJsonLd("my-asset-1", Map.of(
-                Prop.Dcterms.TITLE, "My Asset"
-            ));
+            var assetJsonLd = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset")
+                    .title("My Asset")
+                    .build()
+            );
 
             createConnector(dsl, today, "https://my-connector/api/dsp");
             createDataOffer(dsl, today, "https://my-connector/api/dsp", assetJsonLd);
@@ -178,8 +190,8 @@ class CatalogApiTest {
             assertThat(dataOfferResult.getConnectorEndpoint()).isEqualTo("https://my-connector/api/dsp");
             assertThat(dataOfferResult.getConnectorOfflineSinceOrLastUpdatedAt()).isEqualTo(today);
             assertThat(dataOfferResult.getConnectorOnlineStatus()).isEqualTo(CatalogDataOffer.ConnectorOnlineStatusEnum.ONLINE);
-            assertThat(dataOfferResult.getAssetId()).isEqualTo("my-asset-1");
-            assertThat(dataOfferResult.getAsset().getAssetId()).isEqualTo("my-asset-1");
+            assertThat(dataOfferResult.getAssetId()).isEqualTo("my-asset");
+            assertThat(dataOfferResult.getAsset().getAssetId()).isEqualTo("my-asset");
             assertThat(dataOfferResult.getAsset().getTitle()).isEqualTo("My Asset");
             assertThat(dataOfferResult.getCreatedAt()).isEqualTo(today.minusDays(5));
         });
@@ -213,30 +225,42 @@ class CatalogApiTest {
             // arrange
             var today = OffsetDateTime.now().withNano(0);
 
-            var assetJsonLd1 = getAssetJsonLd("my-asset-1", Map.of(
-                Prop.Mds.DATA_CATEGORY, "my-category-1",
-                Prop.Mds.TRANSPORT_MODE, "MY-TRANSPORT-MODE-1",
-                Prop.Mds.DATA_SUBCATEGORY, "MY-SUBCATEGORY-2",
-                Prop.Mds.DATA_MODEL, "my-data-model",
-                Prop.Mds.GEO_REFERENCE_METHOD, "my-geo-ref"
-            ));
+            var assetJsonLd1 = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset-1")
+                    .dataCategory("my-category-1")
+                    .transportMode("MY-TRANSPORT-MODE-1")
+                    .dataSubcategory("MY-SUBCATEGORY-2")
+                    .dataModel("my-data-model")
+                    .geoReferenceMethod("my-geo-ref")
+                    .build()
+            );
 
-            var assetJsonLd2 = getAssetJsonLd("my-asset-2", Map.of(
-                Prop.Mds.DATA_CATEGORY, "my-category-1",
-                Prop.Mds.TRANSPORT_MODE, "my-transport-mode-2",
-                Prop.Mds.DATA_SUBCATEGORY, "MY-SUBCATEGORY-2"
-            ));
+            var assetJsonLd2 = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset-2")
+                    .dataCategory("my-category-1")
+                    .transportMode("my-transport-mode-2")
+                    .dataSubcategory("MY-SUBCATEGORY-2")
+                    .build()
+            );
 
-            var assetJsonLd3 = getAssetJsonLd("my-asset-3", Map.of(
-                Prop.Mds.DATA_CATEGORY, "my-category-1",
-                Prop.Mds.TRANSPORT_MODE, "MY-TRANSPORT-MODE-1",
-                Prop.Mds.DATA_SUBCATEGORY, "my-subcategory-1"
-            ));
+            var assetJsonLd3 = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset-3")
+                    .dataCategory("my-category-1")
+                    .transportMode("MY-TRANSPORT-MODE-1")
+                    .dataSubcategory("my-subcategory-1")
+                    .build()
+            );
 
-            var assetJsonLd4 = getAssetJsonLd("my-asset-4", Map.of(
-                Prop.Mds.DATA_CATEGORY, "my-category-1",
-                Prop.Mds.TRANSPORT_MODE, ""
-            ));
+            var assetJsonLd4 = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset-4")
+                    .dataCategory("my-category-1")
+                    .transportMode("")
+                    .build()
+            );
 
             createOrganizationMetadata(dsl, "MDSL123456AA", "Test Org");
             createConnector(dsl, today, "https://my-connector/api/dsp", "MDSL123456AA");
@@ -248,32 +272,32 @@ class CatalogApiTest {
             var result = brokerServerClient().brokerServerApi().catalogPage(new CatalogPageQuery());
 
             assertThat(result.getAvailableFilters().getFields())
-                    .extracting(CnfFilterAttribute::getId)
-                    .containsExactly(
-                            "dataSpace",
-                            "dataCategory",
-                            "dataSubcategory",
-                            "dataModel",
-                            "transportMode",
-                            "geoReferenceMethod",
-                            "curatorOrganizationName",
-                            "curatorMdsId",
-                            "connectorEndpoint"
-                    );
+                .extracting(CnfFilterAttribute::getId)
+                .containsExactly(
+                    "dataSpace",
+                    "dataCategory",
+                    "dataSubcategory",
+                    "dataModel",
+                    "transportMode",
+                    "geoReferenceMethod",
+                    "curatorOrganizationName",
+                    "curatorMdsId",
+                    "connectorEndpoint"
+                );
 
             assertThat(result.getAvailableFilters().getFields())
-                    .extracting(CnfFilterAttribute::getTitle)
-                    .containsExactly(
-                            "Data Space",
-                            "Data Category",
-                            "Data Subcategory",
-                            "Data Model",
-                            "Transport Mode",
-                            "Geo Reference Method",
-                            "Organization Name",
-                            "MDS ID",
-                            "Connector"
-                    );
+                .extracting(CnfFilterAttribute::getTitle)
+                .containsExactly(
+                    "Data Space",
+                    "Data Category",
+                    "Data Subcategory",
+                    "Data Model",
+                    "Transport Mode",
+                    "Geo Reference Method",
+                    "Organization Name",
+                    "MDS ID",
+                    "Connector"
+                );
 
             var dataSpace = getAvailableFilter(result, "dataSpace");
             assertThat(dataSpace.getValues()).extracting(CnfFilterItem::getId).containsExactly("MDS");
@@ -325,7 +349,12 @@ class CatalogApiTest {
             // arrange
             var today = OffsetDateTime.now().withNano(0);
 
-            var assetJsonLd = getAssetJsonLd("123", Map.of(Prop.Dcterms.TITLE, "Hello"));
+            var assetJsonLd = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("123")
+                    .title("Hello")
+                    .build()
+            );
 
             createConnector(dsl, today, "https://my-connector/api/dsp");
             createDataOffer(dsl, today, "https://my-connector/api/dsp", assetJsonLd);
@@ -343,8 +372,8 @@ class CatalogApiTest {
 
     private CnfFilterAttribute getAvailableFilter(CatalogPageResult result, String filterId) {
         return result.getAvailableFilters().getFields().stream()
-                .filter(it -> it.getId().equals(filterId)).findFirst()
-                .orElseThrow(() -> new IllegalStateException("Filter not found"));
+            .filter(it -> it.getId().equals(filterId)).findFirst()
+            .orElseThrow(() -> new IllegalStateException("Filter not found"));
     }
 
     @Test
@@ -353,14 +382,20 @@ class CatalogApiTest {
             // arrange
             var today = OffsetDateTime.now().withNano(0);
 
-            var assetJsonLd1 = getAssetJsonLd("my-asset-1", Map.of(
-                Prop.Mds.DATA_CATEGORY, "my-category",
-                Prop.Mds.DATA_SUBCATEGORY, "my-subcategory"
-            ));
+            var assetJsonLd1 = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset-1")
+                    .dataCategory("my-category")
+                    .dataSubcategory("my-subcategory")
+                    .build()
+            );
 
-            var assetJsonLd2 = getAssetJsonLd("my-asset-2", Map.of(
-                Prop.Mds.DATA_SUBCATEGORY, "my-other-subcategory"
-            ));
+            var assetJsonLd2 = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id("my-asset-2")
+                    .dataSubcategory("my-other-subcategory")
+                    .build()
+            );
 
             createConnector(dsl, today, "https://my-connector/api/dsp");
             createDataOffer(dsl, today, "https://my-connector/api/dsp", assetJsonLd1);
@@ -368,7 +403,7 @@ class CatalogApiTest {
 
             var query = new CatalogPageQuery();
             query.setFilter(new CnfFilterValue(List.of(
-                    new CnfFilterValueAttribute("dataCategory", List.of(""))
+                new CnfFilterValueAttribute("dataCategory", List.of(""))
             )));
 
             var result = brokerServerClient().brokerServerApi().catalogPage(query);
@@ -399,7 +434,7 @@ class CatalogApiTest {
 
             var result = brokerServerClient().brokerServerApi().catalogPage(query);
             assertThat(result.getDataOffers()).extracting(CatalogDataOffer::getAssetId)
-                    .isEqualTo(range(0, 10).mapToObj("my-asset-%d"::formatted).toList());
+                .isEqualTo(range(0, 10).mapToObj("my-asset-%d"::formatted).toList());
 
             var actual = result.getPaginationMetadata();
             assertThat(actual.getPageOneBased()).isEqualTo(1);
@@ -428,7 +463,7 @@ class CatalogApiTest {
             var result = brokerServerClient().brokerServerApi().catalogPage(query);
 
             assertThat(result.getDataOffers()).extracting(CatalogDataOffer::getAssetId)
-                    .isEqualTo(range(10, 15).mapToObj("my-asset-%d"::formatted).toList());
+                .isEqualTo(range(10, 15).mapToObj("my-asset-%d"::formatted).toList());
 
             var actual = result.getPaginationMetadata();
             assertThat(actual.getPageOneBased()).isEqualTo(2);
@@ -459,9 +494,9 @@ class CatalogApiTest {
 
             var result = brokerServerClient().brokerServerApi().catalogPage(query);
             assertThat(result.getDataOffers()).extracting(CatalogDataOffer::getAssetId).containsExactly(
-                    "asset-2",
-                    "asset-1",
-                    "asset-3"
+                "asset-2",
+                "asset-1",
+                "asset-3"
             );
         });
     }
@@ -568,7 +603,7 @@ class CatalogApiTest {
             // act
             var query = new CatalogPageQuery();
             query.setFilter(new CnfFilterValue(List.of(
-                    new CnfFilterValueAttribute("curatorOrganizationName", List.of("Unknown"))
+                new CnfFilterValueAttribute("curatorOrganizationName", List.of("Unknown"))
             )));
 
             var actual = brokerServerClient().brokerServerApi().catalogPage(query);
@@ -650,15 +685,15 @@ class CatalogApiTest {
 
     private Policy dummyPolicy() {
         return Policy.Builder.newInstance()
-                .type(PolicyType.SET)
-                .build();
+            .type(PolicyType.SET)
+            .build();
     }
 
     private DataOfferDetailPageResult dataOfferDetails(String endpoint, String assetId) {
         var query = DataOfferDetailPageQuery.builder()
-                .connectorEndpoint(endpoint)
-                .assetId(assetId)
-                .build();
+            .connectorEndpoint(endpoint)
+            .assetId(assetId)
+            .build();
         return brokerServerClient().brokerServerApi().dataOfferDetailPage(query);
     }