+++ extensions/broker-server/src/test/java/de/sovity/edc/ext/brokerserver/services/refreshing/offers/DataOfferWriterTestDataHelper.java
@@ -20,8 +20,8 @@ import de.sovity.edc.ext.brokerserver.db.jooq.tables.records.DataOfferRecord;
 import de.sovity.edc.ext.brokerserver.services.ConnectorCreator;
 import de.sovity.edc.ext.brokerserver.services.refreshing.offers.model.FetchedContractOffer;
 import de.sovity.edc.ext.brokerserver.services.refreshing.offers.model.FetchedDataOffer;
+import de.sovity.edc.ext.wrapper.api.common.model.UiAssetCreateRequest;
 import de.sovity.edc.utils.JsonUtils;
-import de.sovity.edc.utils.jsonld.vocab.Prop;
 import org.apache.commons.lang3.Validate;
 import org.jetbrains.annotations.NotNull;
 import org.jooq.DSLContext;
@@ -30,7 +30,6 @@ import org.jooq.JSONB;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -115,7 +114,12 @@ class DataOfferWriterTestDataHelper {
     }
 
     public String dummyAssetJson(Do dataOffer) {
-        var assetJsonLd = getAssetJsonLd(dataOffer.getAssetId(), Map.of(Prop.Dcterms.TITLE, dataOffer.getAssetTitle()));
+        var assetJsonLd = getAssetJsonLd(
+                UiAssetCreateRequest.builder()
+                    .id(dataOffer.getAssetId())
+                    .title(dataOffer.getAssetTitle())
+                    .build()
+        );
         return JsonUtils.toJson(assetJsonLd);
     }